Exceptionhandling for SA and DAL added

+++ src/DataAccess/SKSGroupF.SKS.Package.DataAccess.Sql/SqlParcelRepository.cs
@@ -1,6 +1,8 @@
 ﻿using Microsoft.Extensions.Logging;
 using SKSGroupF.SKS.Package.DataAccess.Entities.Models;
 using SKSGroupF.SKS.Package.DataAccess.Interfaces;
+using SKSGroupF.SKS.Package.DataAccess.Interfaces.Exceptions;
+using System;
 using System.Collections.Generic;
 using System.Linq;
 
@@ -19,12 +21,32 @@ namespace SKSGroupF.SKS.Package.DataAccess.Sql
 
         public int Create(DALParcel parcel)
         {
-            //Insert Parcel and return db ID
-            logger.LogInformation("Inserting parcel into database.");
-            var ent = context.DbParcel.Add(parcel);
-            logger.LogInformation("Parcel inserted with ID + " + ent.Entity.Id + ".");
-            SaveChanges();
-            return ent.Entity.Id;
+            try
+            {
+                try
+                {
+                    GetByTrackingId(parcel.TrackingId);
+                }
+                catch (DALDataNotFoundException) //Parcel will only be created if GetByTrackingId does not find a parcel with this tracking Id in the database
+                {
+                    //Insert Parcel and return db ID
+                    logger.LogInformation("Inserting parcel into database.");
+                    var ent = context.DbParcel.Add(parcel);
+                    logger.LogInformation("Parcel inserted with ID + " + ent.Entity.Id + ".");
+                    SaveChanges();
+                    return ent.Entity.Id;
+                }
+
+                string errorMsg = "Parcel with specified tracking Id does already exsist in database.";
+                logger.LogError(errorMsg);
+                throw new DALDataDuplicateException(nameof(SqlParcelRepository), nameof(Create), errorMsg);
+            }
+            catch (Exception ex)
+            {
+                string errorMsg = "There has been an error with the database";
+                logger.LogError(errorMsg);
+                throw new DALDataException(nameof(SqlParcelRepository), nameof(Create), errorMsg, ex);
+            }           
         }
 
         public void Update(DALParcel parcel)
@@ -38,6 +60,12 @@ namespace SKSGroupF.SKS.Package.DataAccess.Sql
         {
             logger.LogInformation("Removing existing parcel from database.");
             var parcel = context.DbParcel.Find(id);
+            if(parcel == null)
+            {
+                string errorMsg = "Parcel with specified Id could not be found in database.";
+                logger.LogError(errorMsg);
+                throw new DALDataNotFoundException(nameof(SqlParcelRepository), nameof(Delete), errorMsg);
+            }
             context.DbParcel.Remove(parcel);
             SaveChanges();
         }
@@ -46,9 +74,15 @@ namespace SKSGroupF.SKS.Package.DataAccess.Sql
         {
             try
             {
+                logger.LogInformation("Trying to select all parcels from database.");
                 return context.DbParcel.ToList();
             }
-            catch { return null; }
+            catch (Exception ex)
+            {
+                string errorMsg = "No parcels found in Database.";
+                logger.LogError(errorMsg);
+                throw new DALDataNotFoundException(nameof(SqlParcelRepository), nameof(GetAll), errorMsg, ex);
+            }
         }
 
         public DALParcel GetByTrackingId(string tid)
@@ -58,14 +92,15 @@ namespace SKSGroupF.SKS.Package.DataAccess.Sql
                 logger.LogInformation("Trying to select a single parcel with tracking Id " + tid + ".");
                 return context.DbParcel.Single(p => p.TrackingId == tid);
             }
-            catch 
+            catch (Exception ex)
             {
-                logger.LogInformation("Failed to select parcel.");
-                return null; 
+                string errorMsg = "Parcel with specified tracking Id could not be found in Database.";
+                logger.LogError(errorMsg);
+                throw new DALDataNotFoundException(nameof(SqlHopRepository), nameof(GetByTrackingId), errorMsg, ex);
             }
         }
 
-        public IEnumerable<DALParcel> GetByReceipient(DALReceipient receipient)
+        /*public IEnumerable<DALParcel> GetByReceipient(DALReceipient receipient)
         {
             try
             {
@@ -99,27 +134,45 @@ namespace SKSGroupF.SKS.Package.DataAccess.Sql
                 return context.DbParcel.Where(p => p.Weight >= min && p.Weight <= max).ToList();
             }
             catch { return null; }
-        }
+        }*/
 
         public void UpdateHopState(DALParcel parcel, string code)
         {
-            logger.LogInformation("Changing hop-state of parcel.");
-            //?
-            var hop = context.DbHopArrival.Single(p => p.Code == code);
-            parcel.FutureHops.Remove(hop);
-            parcel.VisitedHops.Add(hop);
-            context.DbParcel.Update(parcel);
-            SaveChanges();
+            try
+            {
+                logger.LogInformation("Changing hop-state of parcel.");
+                //?
+                var hop = context.DbHopArrival.Single(p => p.Code == code);
+                parcel.FutureHops.Remove(hop);
+                parcel.VisitedHops.Add(hop);
+                context.DbParcel.Update(parcel);
+                SaveChanges();
+            }
+            catch (Exception ex)
+            {
+                string errorMsg = "Could not update hop-state.";
+                logger.LogError(errorMsg);
+                throw new DALDataException(nameof(SqlHopRepository), nameof(GetAll), errorMsg, ex);
+            }
             //Selects FutureHop by Code and marks it as VisitedHop
         }
 
         public void UpdateDelivered(DALParcel parcel)
         {
-            logger.LogInformation("Changing delivered-state of parcel.");
-            parcel.Delievered = true;
-            context.DbParcel.Update(parcel);
-            SaveChanges();
-            //Sets parcel as delievered ??
+            try
+            {
+                logger.LogInformation("Changing delivered-state of parcel.");
+                parcel.Delievered = true;
+                context.DbParcel.Update(parcel);
+                SaveChanges();
+                //Sets parcel as delievered ??
+            }
+            catch (Exception ex)
+            {
+                string errorMsg = "Could not update delivered-state state.";
+                logger.LogError(errorMsg);
+                throw new DALDataException(nameof(SqlHopRepository), nameof(GetAll), errorMsg, ex);
+            }
         }
 
         int SaveChanges()
